bugfix: change path.join to explicitly use path.posix.join when passed to globby
I have changed each instance of  to explicitly use  when that path is passed into globby. This is to fix an issue where, on windows platforms, a windows path is passed into globby. This is not allowed according to globby's documentation and was causing all directories in  to not load correctly on the website.

Co-authored-by: GDWR <gregory.dwr@gmail.com>